// src/app/model/roles.model.ts
export class Roles{
    roleId:number;
    roleName: string;
    roleStatus: string;
    roleDescription: string;
    // permission: Permission;
    // menuItems: MenuItem[];
    createdBy: string;
    createdByEmailId: string;
    createdDateTime : Date;
    modifiedBy: string;
    modifiedByEmailId: string;
    modifiedDateTime: Date;



    // /**
    //  *
    //  */
    // constructor() {
    //     super();
    // }

}